requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database (Supabase/Postgres)
supabase>=2.3.0
asyncpg>=0.29.0

# Stripe billing
stripe>=7.10.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
